refactor(os-exporter): move dto
Moves the dto from the test classes to the dto classes. This way we can
reuse it.
